Move all first class citizens into base.py
